// block-performance-1.0.0/src/library/cloudman/scp-download/type.ts
export interface Input {
    remote_ssh_auth: { username: string; address: string; ssh_auth_key: { os?: boolean; store?: boolean; file?: string } };
    from_remote_path: string;
    to_local_path: string;
}

export interface OutputOk {}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}